chore(apps/interface): create insight generator component
